requirements.txt
Django==5.2.7

# REST
 djangorestframework==3.16.1
 djangorestframework-simplejwt==5.3.1
 drf-spectacular==0.27.2

# CORS
 django-cors-headers==4.9.0

# Filters
 django-filter==25.2

# Google Places API
googlemaps==4.10.0

# Environment variables
python-decouple==3.8

# HTTP requests
requests==2.31.0

# Database
mongoengine==0.29.1
pymongo==4.15.3

# Optional: MongoDB support (uncomment if using Mongo)
# djongo==1.3.6
# pymongo==4.10.1
